requirements.txt
streamlit
fastapi
Pillow
